requirements.txt
Pillow==8.2.0
base58==2.1.0
imageio==2.9.0
matplotlib==3.3.4
pandas==1.2.4
streamlit==0.84.1
requests==2.25.1
retrying==1.3.3
imageio==2.9.0
lxml==4.6.3
openpyxl==3.0.7
packaging==20.9
pytrends==4.7.3
plotly==4.14.3
imageio==2.9.0
